@@ -5,51 +5,53 @@
     "license": "MIT",
     "type": "project",
     "require": {
-        "php": "7.1.*",
-        "laravel/framework": "5.2.*",
-        "dingo/api": "1.0.x@dev",
-        "tymon/jwt-auth": "^0.5.12",
-        "barryvdh/laravel-cors":"^0.9.2",
+        "php": "^7.2",
         "ckeditor/ckeditor": "^4.7",
-        "doctrine/dbal": "^2.6"
+        "dingo/api": "~2.4.7",
+        "doctrine/dbal": "^2.10.2",
+        "fideloper/proxy": "^4.4.0",
+        "fruitcake/laravel-cors": "^2.0.1",
+        "laravel/framework": "~6.18.22",
+        "laravel/tinker": "^2.4.0",
+        "tymon/jwt-auth": "^1.0"
     },
     "require-dev": {
-        "fzaninotto/faker": "~1.4",
-        "mockery/mockery": "0.9.*",
-        "phpunit/phpunit": "~4.0",
-        "symfony/css-selector": "2.8.*|3.0.*",
-        "symfony/dom-crawler": "2.8.*|3.0.*"
+        "symfony/css-selector": "~5.1.2",
+        "symfony/dom-crawler": "~5.1.2",
+        "filp/whoops": "~2.7.3"
     },
     "autoload": {
         "classmap": [
-            "database"
+            "database/seeds",
+            "database/factories"
         ],
         "psr-4": {
             "App\\": "app/"
         }
     },
-    "autoload-dev": {
-        "classmap": [
-            "tests/TestCase.php"
-        ]
+    "autoload-dev": { },
+    "config": {
+        "optimize-autoloader": true,
+        "preferred-install": "dist",
+        "sort-packages": true
     },
+    "extra": {
+        "laravel": {
+            "dont-discover": []
+        }
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true,
     "scripts": {
+        "post-autoload-dump": [
+            "Illuminate\\Foundation\\ComposerScripts::postAutoloadDump",
+            "@php artisan package:discover --ansi"
+        ],
         "post-root-package-install": [
-            "php -r \"copy('.env.example', '.env');\""
+            "@php -r \"file_exists('.env') || copy('.env.example', '.env');\""
         ],
         "post-create-project-cmd": [
-            "php artisan key:generate"
-        ],
-        "post-install-cmd": [
-            "Illuminate\\Foundation\\ComposerScripts::postInstall",
-            "php artisan optimize"
-        ],
-        "post-update-cmd": [
-            "Illuminate\\Foundation\\ComposerScripts::postUpdate",
-            "php artisan optimize"
+            "@php artisan key:generate --ansi"
         ]
-    },
-    "config": {
-        "preferred-install": "dist"
     }
 }
