@@ -3,15 +3,16 @@
 namespace App\Http\APIControllers;
 
 use App\Application;
-use App\ClassOfStudents;
+use App\Mail\PasswordResetMail;
 use Illuminate\Http\Request;
-use JWTAuth;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Str;
+use Tymon\JWTAuth\Facades\JWTAuth;
 use Tymon\JWTAuth\Exceptions\JWTException;
 use Illuminate\Support\Facades\DB;
-use Validator;
 use App\Student;
 use App\PasswordResets;
-use Mail;
+use Illuminate\Support\Facades\Mail;
 
 class AuthController extends Controller
 {
@@ -19,21 +20,17 @@ class AuthController extends Controller
     public function authenticate(Request $request)
     {
         auth()->shouldUse('api');
-        // grab credentials from the request
         $credentials = $request->only('email', 'password');
         try {
-            // attempt to verify the credentials and create a token for the user
             if (! $token = JWTAuth::attempt($credentials)) {
                 return $this->error('invalid_credentials');
             }
         } catch (JWTException $e) {
-            // something went wrong whilst attempting to encode the token
-            throw new \Symfony\Component\HttpKernel\Exception\HttpException(500, 'could_not_create_token');
+            throw new \Symfony\Component\HttpKernel\Exception\HttpException(500, 'Could Not Create Token');
         }
-        //if user is admin update corresponding field
-        $student_id = JWTAuth::getPayload($token)->get('sub');
-        DB::unprepared("UPDATE students s LEFT JOIN users u ON s.email = u.email SET s.is_admin = IF(u.id, 1, 0) WHERE s.id = ".$student_id);
-        $student = Student::find($student_id);
+        $student = auth()->user();
+        DB::unprepared("UPDATE students s LEFT JOIN users u ON s.email = u.email SET s.is_admin = IF(u.id, 1, 0) WHERE s.id = ".$student->id);
+
         $student->question_num = $student->question_num ?: 5;
         if (!$student->app_id) {
             $student->app_id = Application::first()->id;
@@ -59,7 +56,6 @@ class AuthController extends Controller
     public function register(Request $request)
     {
         $fields = ['email', 'password', 'name'];
-        // grab credentials from the request
         $credentials = $request->only($fields);
         foreach($fields as $field) {
             $credentials[$field] = trim($credentials[$field]);
@@ -85,25 +81,19 @@ class AuthController extends Controller
         if ($result) {
             return $this->success($result);
         }
-        return $this->success($error);
+        return $this->error('Something went wrong!');
     }
 
     public function passwordResetEmail(Request $request) {
         $fields = ['email', 'url'];
-        // grab credentials from the request
         $credentials = $request->only($fields);
         foreach($fields as $field) {
             $credentials[$field] = trim($credentials[$field]);
         }
-        $validator = Validator::make(
-            $credentials,
-            [
-                'email' => 'required|email|max:255',
-                'url' => 'required'
-            ]
-            );
-        if ($validator->fails())
-        {
+        $validator = Validator::make($credentials,
+            [ 'email' => 'required|email|max:255', 'url' => 'required' ]
+        );
+        if ($validator->fails()) {
             return $this->error($validator->messages());
         }
         $email = $credentials['email'];
@@ -111,62 +101,49 @@ class AuthController extends Controller
         if(!$student) {
             return $this->error('We can not find email you provided in our database! You can register a new account with this email.');
         }
-        // delete existings resets if exists
         PasswordResets::where('email', $email)->delete();
-        $url = $credentials['url'];
-        $token = str_random(64);
+        $token = Str::random(64);
         $result = PasswordResets::create([
             'email' => $email,
             'token' => $token
         ]);
-        if($result) {
-            $data = array('token'=>$token, 'email' => $email, 'url' => $url);
-            Mail::queue('mail', $data, function($message) use ($data) {
-               $message->to($data['email'], $data['email'])->subject
-                  ('Password Reset at MathGizmo');
-               $message->from('mathgizmo01@gmail.com', 'Gizmo Support');
-            });
+        if ($result) {
+            Mail::to($email)->send(new PasswordResetMail($credentials['url'] . '/' . $token));
             return $this->success('The mail has been sent successfully!');
         }
-        return $this->success($error);
+        return $this->error('Something went wrong!');
     }
 
     public function resetPassword(Request $request) {
         $fields = ['password', 'token'];
-        // grab credentials from the request
         $credentials = $request->only($fields);
         foreach($fields as $field) {
             $credentials[$field] = trim($credentials[$field]);
         }
-        $validator = Validator::make(
-            $credentials,
-            [
-                'password' => 'required|min:6',
-                'token' => 'required'
-            ]
-            );
-        if ($validator->fails())
-        {
+        $validator = Validator::make($credentials,
+            [ 'password' => 'required|min:6', 'token' => 'required' ]
+        );
+        if ($validator->fails()) {
             return $this->error($validator->messages());
         }
         $token = $credentials['token'];
         $pr = PasswordResets::where('token', $token)->first(['email', 'created_at']);
         $email = $pr['email'];
-        if(!$email) {
+        if (!$email) {
             return $this->error('Invalid reset password link!');
         }
         $dateCreated = strtotime($pr['created_at']);
         $expireInterval = 86400; // token expire interval in seconds (24 h)
         $currentTime = time();
-        if($currentTime  - $dateCreated > $expireInterval) {
+        if ($currentTime - $dateCreated > $expireInterval) {
             return $this->error('The time to reset password has expired!');
         }
         $password = bcrypt($credentials['password']);
         $updatedRows = Student::where('email', $email)->update(['password' => $password]);
-        if($updatedRows > 0) {
+        if ($updatedRows > 0) {
             PasswordResets::where('token', $token)->delete();
             return $this->success('The password has been changed successfully!');
         }
-        return $this->success($error);
+        return $this->error('Something went wrong!');
     }
 }
