@@ -16,7 +16,7 @@ class FileController extends Controller
             $request->file('icon')->move('images/icons', $new_name);
             //$request->file('icon')->move('images/icons', $new_name_complete);
             $json['path'] = 'images/icons/'.$new_name;
-        } catch (Exception $e) {
+        } catch (\Exception $e) {
             $json['path'] = 'Caught exception: '.$e->getMessage();
         }
         return json_encode($json);
@@ -28,15 +28,15 @@ class FileController extends Controller
         try {
             if($request->icon) {
                 if(strpos($request->icon, 'images/icons/') === 0) {
-                   unlink($request->icon); 
+                   unlink($request->icon);
                 }
                 /*$complete_icon = str_replace(".svg","-gold.svg",$request->icon);
                 if(strpos($complete_icon, 'images/icons/') === 0) {
                     unlink($complete_icon);
                 }*/
-            } 
+            }
             $json['icon'] = $request->icon;
-        } catch (Exception $e) {
+        } catch (\Exception $e) {
             $json['icon'] = 'Caught exception: '.$e->getMessage();
         }
         return json_encode($json);
