@@ -7,14 +7,7 @@ use Illuminate\Support\Facades\Auth;
 
 class Authenticate
 {
-    /**
-     * Handle an incoming request.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \Closure  $next
-     * @param  string|null  $guard
-     * @return mixed
-     */
+
     public function handle($request, Closure $next, $guard = null)
     {
         if (Auth::guard($guard)->guest()) {
@@ -24,7 +17,6 @@ class Authenticate
                 return redirect()->guest('login');
             }
         }
-
         return $next($request);
     }
 }
