@@ -2,9 +2,9 @@
 
 namespace App\Providers;
 
+use App\ClassOfStudents;
+use App\Observers\ClassObserver;
 use Illuminate\Support\ServiceProvider;
-use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Log;
 
 class AppServiceProvider extends ServiceProvider
 {
@@ -15,13 +15,7 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        //degug mode
-        /*DB::listen(function ($query) {
-            // $query->sql
-            // $query->bindings
-            // $query->time
-            Log::info("Executed:".$query->sql);
-        });*/
+        ClassOfStudents::observe(ClassObserver::class);
     }
 
     /**
