@@ -11,6 +11,8 @@ return [
     | framework when an event needs to be broadcast. You may set this to
     | any of the connections defined in the "connections" array below.
     |
+    | Supported: "pusher", "redis", "log", "null"
+    |
     */
 
     'default' => env('BROADCAST_DRIVER', 'pusher'),
@@ -30,11 +32,15 @@ return [
 
         'pusher' => [
             'driver' => 'pusher',
-            'key' => env('PUSHER_KEY'),
-            'secret' => env('PUSHER_SECRET'),
-            'app_id' => env('PUSHER_APP_ID'),
+            'key' => env('PUSHER_APP_KEY', 'gizmo'),
+            'secret' => env('PUSHER_APP_SECRET', 'gizmo'),
+            'app_id' => env('PUSHER_APP_ID', 'gizmo'),
             'options' => [
-                //
+                'cluster' => null,
+                'encrypted' => false,
+                'host' => '127.0.0.1',
+                'port' => 6001,
+                'scheme' => 'http'
             ],
         ],
 
@@ -45,6 +51,10 @@ return [
 
         'log' => [
             'driver' => 'log',
+        ],
+
+        'null' => [
+            'driver' => 'null',
         ],
 
     ],
